backend/requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
python-chess==1.10.0
python-multipart==0.0.6
requests==2.31.0
